Add delay to mouse wheel event from drag start
This adds a 1/5 of a second delay from the registration of a drag event
to actually starting to scroll. The result is a slight decrease in
sensitivity when dragging a finger on the Magic Mouse to start a scroll.
The actual scrolling is the same as before, there is a just a small
delay at the beginning of the stroke which helps to eliminate accidental
scrolls when just resting the finger lightly on the mouse surface.