Improve Driver Loading through Modprobe and DKMS
This update simplifies the code. Please run `remove.sh` before updating the driver to ensure a smooth transition.

Removed unnecessary files from repo.

Changes were made to install, the main loading script to make use of modprobe more accurately.

The parameters for hid-magicmouse were moved to /etc/modprobe.d/hid-magicmouse.conf . This allows the unloading and loading of the module with predefined parameters.

A `remove.sh` script was added to uninstall the driver. NOTE: It does not remove the bluetooth fix.

Changed DKMS version to reflect the driver name.